[quant][graphmode][fx][refactor] Move patterns to separate files
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
